Wrapper to encode/decode token
Given payload is encoded assuming some default claims to be present:

- iat: Issued At Time. Always takes current time.
- exp: Expiration Time. Configurable in number of seconds from now in
  `expiration_time` setting
- jti: JWT id. Unique identifier

For now, HS256 algorithm is always used. Secret must be configured in
`secret` setting.

A JWT blacklist can be configured in a `blacklist` setting. This can be
anything responding a boolean to an `include?` method. If given token
`jti` is included in the blacklist decoding is not performed.