Warden strategy for JWT authentication
The strategy is valid if `Authorization` header is present

It takes a JWT token from an `Authorization` header that is expected to
have a `Bearer #{token}` form.

If the token can't be decoded it fails and halts the strategies cascade.

It the token can be decoded, `sub` (subject) claim is taken and passed
as an argument to a `find_for_jwt_authentication` method to a configured
mapping for current warden scope.

Mappings can be configured in `mappings` settings, and they are a hash
{ warden_scope: mapping }