Middleware to add token to the response
URLs for which add the token are specified as a Regex in
`config.response_token_paths`. Token is added in `Authorization`
response header, in the form "Bearer #{token}"

Actual middleware is called through a 'dispatcher' middleware because
another one will be added to control when token is added to the
blacklist.