Do not run children middleware if user is not logged in
They are the ones that work once the actual application has been reached
and it has logged in or not an user.

If an user has not been logged in, there is no way to create a token for
it. Also, if an user has not been logged in, it is because there is no
token in the headers or maybe it is expired, so there is no need to add
it to the blacklist.