Abstract to any revocation strategy
Blacklisting is not the only JWT revocation strategy that can be
implemented, so it is better to abstract it.

Configured revocation strategy should respond to two methods:

* `revoke(payload)`, which does something to persist the state about that
 token being revoked

* `revoke?(payload)`, which should return a boolean whether a given token
 has been revoked

I prefer to pass the payload as argument instead of the whole token,
because it enforces that the whole token won't be persisted (that would
be dangerous if some security concerns are not taken into account, like
encrypting and salting the tokens).