TokenCoder takes any payload as argument
This is a first step to add a user coder layer on top of TokenCoder
which will work with the actual user to encode/decode. This way,
TokenCoder is left agnostic about what has to be encoded (in the same
way it was already agnostic about what was being decoded).