Allow adding info from the user to the token on creation
This makes unnecessary the hook method that was called in the user at
token dispatch time. The intention of that hook method was to update the
user with some info in the token to be used later on revocation check
time. Now the same can be achieved reading from the user and copying to
the token on creation time.