Encode user scope in the payload
To do so, token creation has been moved from a rack middleware to a
warden callback.

It is cleaner to get the scope from within the callback (it is provided
in the arguments) that in the middleware (where accessing `@config`
private instance variable in the warden proxy object was needed).

The additional check of whether the scope is known by
Warden::JWTAuth.config has been included, so that now another unknown
scoped also logged in will not interfere. However, still it is only
possible to dispatch a single JWT token in the `Authorization` header
(which would belong to the last known scope user who signs in). It would
be that difficult to extend that and, say, separate tokens by `,` but it
is not an accepted standard, as far as I'm concerned.